Add Stb2Brep to Bake
